Changes based on comments
