+++ src/main/java/net/sf/jabref/importer/AutosaveStartupPrompter.java
@@ -73,7 +73,7 @@ public class AutosaveStartupPrompter implements Runnable {
             boolean done = false;
             ParserResult pr;
             do {
-                pr = AutosaveAwareDatabaseLoader.openBibFile(fileToLoad.getPath(), true);
+                pr = OpenDatabaseAction.openBibFile(fileToLoad.getPath(), true);
                 if (pr.isInvalid()) {
                     loaded.add(pr);
                     BibDatabaseContext databaseContext = pr.getDatabaseContext();
